+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlayground.swift
@@ -57,7 +57,6 @@ struct PaymentSheetTestPlayground: View {
         SettingView(setting: $playgroundController.settings.shakeAmbiguousViews)
         SettingView(setting: $playgroundController.settings.instantDebitsIncentives)
         SettingView(setting: $playgroundController.settings.fcLiteEnabled)
-        SettingView(setting: $playgroundController.settings.linkInSPMs)
         SettingView(setting: $playgroundController.settings.linkFlowControllerChanges)
     }