+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -592,12 +592,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case allowVisa
     }
 
-    enum LinkInSPMs: String, PickerEnum {
-        static let enumName: String = "Link in SPM"
-        case on
-        case off
-    }
-
     enum LinkFlowControllerChanges: String, PickerEnum {
         static let enumName: String = "Link FlowController RUX"
         case on
@@ -664,7 +658,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
     var embeddedViewDisplaysMandateText: DisplaysMandateTextEnabled
     var rowSelectionBehavior: RowSelectionBehavior
     var cardBrandAcceptance: CardBrandAcceptance
-    var linkInSPMs: LinkInSPMs
     var linkFlowControllerChanges: LinkFlowControllerChanges
 
     static func defaultValues() -> PaymentSheetTestPlaygroundSettings {
@@ -719,7 +712,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
             embeddedViewDisplaysMandateText: .on,
             rowSelectionBehavior: .default,
             cardBrandAcceptance: .all,
-            linkInSPMs: .off,
             linkFlowControllerChanges: .off
         )
     }