+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/Link/LinkUI.swift
@@ -35,7 +35,24 @@ enum LinkUI {
 
     // MARK: - Border
 
-    static let borderWidth: CGFloat = 1.5
+    static let borderWidth: CGFloat = 1.0
+
+    static let highlightBorderConfiguration = HighlightBorderConfiguration(
+        width: borderWidth,
+        color: UIColor.linkBorderSelected.cgColor,
+        animator: animator
+    )
+
+    // MARK: - Buttons
+
+    private static let minimumLabelHeight: CGFloat = 24
+
+    private static let minimumButtonHeight: CGFloat = 44
+
+    static func primaryButtonHeight(margins: NSDirectionalEdgeInsets) -> CGFloat {
+        let height = LinkUI.minimumLabelHeight + margins.top + margins.bottom
+        return max(height, minimumButtonHeight)
+    }
 
     // MARK: - Margins
 
@@ -56,6 +73,19 @@ enum LinkUI {
     static let smallContentSpacing: CGFloat = 8
 
     static let tinyContentSpacing: CGFloat = 4
+
+    // MARK: - Animations
+
+    static let animator: UIViewPropertyAnimator = {
+        let params = UISpringTimingParameters(
+            mass: 1.0,
+            dampingRatio: 0.93,
+            frequencyResponse: 0.22
+        )
+        let animator = UIViewPropertyAnimator(duration: 0, timingParameters: params)
+        animator.isInterruptible = true
+        return animator
+    }()
 }
 
 // MARK: Development flags