+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/New Payment Method Screen/PaymentMethodTypeCollectionView.swift
@@ -69,8 +69,8 @@ class PaymentMethodTypeCollectionView: UICollectionView {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .horizontal
         layout.sectionInset = UIEdgeInsets(
-            top: 0, left: appearance.formInsets.left, bottom: 0,
-            right: appearance.formInsets.right)
+            top: 0, left: appearance.formInsets.leading, bottom: 0,
+            right: appearance.formInsets.trailing)
         layout.minimumInteritemSpacing = PaymentMethodTypeCollectionView.minInteritemSpacing
         super.init(frame: .zero, collectionViewLayout: layout)
         self.dataSource = self