+++ CHANGELOG.md
@@ -1,5 +1,15 @@
-## XX.XX.X XXXX-XX-XX
+## x.x.x x-x-x
+
 ### PaymentSheet
+* [Fixed] A bug where the EmbeddedPaymentElement failed to display forms when tapped in SwiftUI.
+
+## 24.15.0 2025-06-02
+
+### Connect
+* [CHANGED] Removed `PrivateBetaConnect` annotation and all usages. [Connect SDK](https://docs.stripe.com/connect/get-started-connect-embedded-components?platform=ios) is now GA.
+
+### PaymentSheet
+* [Added] The ability to customize the height of the primary button with `PaymentSheetAppearance.primaryButton.height`.
 * [Added] The ability to customize form insets with `PaymentSheetAppearance.formInsets`.
 
 ## 24.14.0 2025-05-19